Fix SRT(LA) custom adaptive bitrate factor.

+++ Moblin/Srtla/AdaptiveBitrate.swift
@@ -10,11 +10,11 @@ struct StreamStats {
 }
 
 struct AdaptiveBitrateSettings {
-    var packetsInFlight: Int32
+    var packetsInFlight: Int64
     var rttDiffHighFactor: Double
     var rttDiffHighAllowedSpike: Double
-    var rttDiffHighMinDecrease: Int32
-    var pifDiffIncreaseFactor: Int32
+    var rttDiffHighMinDecrease: Int64
+    var pifDiffIncreaseFactor: Int64
 }
 
 let adaptiveBitrateFastSettings = AdaptiveBitrateSettings(
@@ -36,10 +36,10 @@ let adaptiveBitrateSlowSettings = AdaptiveBitrateSettings(
 class AdaptiveBitrate {
     private var avgRtt: Double = 0.0
     private var fastRtt: Double = 0.0
-    private var curBitrate: Int32 = 250_000
-    private var prevBitrate: Int32 = 250_000
-    private var targetBitrate: Int32 = 250_000
-    private var tempMaxBitrate: Int32 = 250_000
+    private var curBitrate: Int64 = 250_000
+    private var prevBitrate: Int64 = 250_000
+    private var targetBitrate: Int64 = 250_000
+    private var tempMaxBitrate: Int64 = 250_000
     private var smoothPif: Double = 0
     private var fastPif: Double = 0
     private weak var delegate: (any AdaptiveBitrateDelegate)!
@@ -47,12 +47,12 @@ class AdaptiveBitrate {
     private var settings = adaptiveBitrateFastSettings
 
     init(targetBitrate: UInt32, delegate: AdaptiveBitrateDelegate) {
-        self.targetBitrate = Int32(targetBitrate)
+        self.targetBitrate = Int64(targetBitrate)
         self.delegate = delegate
     }
 
     func setTargetBitrate(bitrate: UInt32) {
-        targetBitrate = Int32(bitrate)
+        targetBitrate = Int64(bitrate)
     }
 
     func setSettings(settings: AdaptiveBitrateSettings) {
@@ -93,9 +93,9 @@ class AdaptiveBitrate {
         stats: StreamStats,
         pif: Double,
         allowedRttJitter: Double,
-        allowedPifJitter: Int32
+        allowedPifJitter: Int64
     ) {
-        var pifDiffThing = Int32(stats.packetsInFlight - pif)
+        var pifDiffThing = Int64(stats.packetsInFlight - pif)
         if pifDiffThing < 0 {
             pifDiffThing = 0
         }
@@ -104,7 +104,7 @@ class AdaptiveBitrate {
         }
         pifDiffThing = settings.packetsInFlight - pifDiffThing
         if pif < Double(settings.packetsInFlight), fastRtt <= avgRtt + allowedRttJitter {
-            if Int32(stats.packetsInFlight - pif) < allowedPifJitter {
+            if Int64(stats.packetsInFlight - pif) < allowedPifJitter {
                 tempMaxBitrate += (settings.pifDiffIncreaseFactor * pifDiffThing) / settings.packetsInFlight
                 if tempMaxBitrate > targetBitrate {
                     tempMaxBitrate = targetBitrate
@@ -131,10 +131,10 @@ class AdaptiveBitrate {
     private func decreaseMaxRateIfPifIsHigh(
         factor: Double,
         pifMax: Double,
-        minimumDecrease: Int32
+        minimumDecrease: Int64
     ) {
         if smoothPif > pifMax {
-            let newMaxBitrate = Int32(Double(tempMaxBitrate) * factor)
+            let newMaxBitrate = Int64(Double(tempMaxBitrate) * factor)
             let differece = tempMaxBitrate - newMaxBitrate
             if differece < minimumDecrease {
                 tempMaxBitrate -= minimumDecrease
@@ -145,7 +145,7 @@ class AdaptiveBitrate {
                     """
                 )
             } else {
-                tempMaxBitrate = Int32(Double(tempMaxBitrate) * factor)
+                tempMaxBitrate = Int64(Double(tempMaxBitrate) * factor)
                 logAdaptiveAcion(
                     actionTaken: """
                     PIF: decreasing bitrate by \(Int((100 * (1 - factor)).rounded()))%, \
@@ -170,10 +170,10 @@ class AdaptiveBitrate {
     private func decreaseMaxRateIfRttIsHigh(
         factor: Double,
         rttMax: Double,
-        minimumDecrease: Int32
+        minimumDecrease: Int64
     ) {
         if avgRtt > rttMax {
-            let newMaxBitrate = Int32(Double(tempMaxBitrate) * factor)
+            let newMaxBitrate = Int64(Double(tempMaxBitrate) * factor)
             let differece = tempMaxBitrate - newMaxBitrate
 
             if differece < minimumDecrease {
@@ -191,11 +191,11 @@ class AdaptiveBitrate {
         }
     }
 
-    var getCurrentBitrate: Int32 {
+    var getCurrentBitrate: Int64 {
         return curBitrate / 1000
     }
 
-    var getTempMaxBitrate: Int32 {
+    var getTempMaxBitrate: Int64 {
         return tempMaxBitrate / 1000
     }
 
@@ -203,22 +203,22 @@ class AdaptiveBitrate {
         return adaptiveActionsTaken
     }
 
-    var getFastPif: Int32 {
-        return Int32(fastPif)
+    var getFastPif: Int64 {
+        return Int64(fastPif)
     }
 
-    var getSmoothPif: Int32 {
-        return Int32(smoothPif)
+    var getSmoothPif: Int64 {
+        return Int64(smoothPif)
     }
 
     private func decreaseMaxRateIfRttDiffIsHigh(
         _ stats: StreamStats,
         factor: Double,
         rttSpikeAllowed: Double,
-        minimumDecrease: Int32
+        minimumDecrease: Int64
     ) {
         if stats.rttMs > avgRtt + rttSpikeAllowed {
-            let newMaxBitrate = Int32(Double(tempMaxBitrate) * factor)
+            let newMaxBitrate = Int64(Double(tempMaxBitrate) * factor)
             let differece = tempMaxBitrate - newMaxBitrate
             if differece < minimumDecrease {
                 tempMaxBitrate -= minimumDecrease
@@ -242,13 +242,13 @@ class AdaptiveBitrate {
     }
 
     private func calculateCurrentBitrate() {
-        var pifDiffThing = Int32(fastPif) - Int32(smoothPif)
+        var pifDiffThing = Int64(fastPif) - Int64(smoothPif)
         // lazy decrease
         if pifDiffThing > settings.packetsInFlight {
             logAdaptiveAcion(
                 actionTaken: "Lazy dec pifdiff \(pifDiffThing) > limit \(settings.packetsInFlight)"
             )
-            tempMaxBitrate = Int32(Double(tempMaxBitrate) * 0.95)
+            tempMaxBitrate = Int64(Double(tempMaxBitrate) * 0.95)
         }
         if pifDiffThing <= (settings.packetsInFlight / 5) {
             pifDiffThing = 0
@@ -274,7 +274,7 @@ class AdaptiveBitrate {
         var tempBitrate = Int64(tempMaxBitrate)
         tempBitrate *= Int64(pifDiffThing)
         tempBitrate /= Int64(settings.packetsInFlight)
-        curBitrate = Int32(tempBitrate)
+        curBitrate = Int64(tempBitrate)
         if curBitrate < 50000 {
             curBitrate = 50000
         }
